[zircon] Fix clippy warnings in kernel Rust code
This fixes or suppresses many of the clippy warnings that popped up in
the Zircon kernel Rust code as a result of recently fixing the clippy
tooling support to check this code.

Bug: 543505311
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1746889
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: James Robinson <jamesr@google.com>

+++ zircon/kernel/lib/debuglog/debuglog.rs
@@ -9,8 +9,8 @@ use zx_status::Status;
 
 #[allow(unused_imports)]
 pub use debuglog_types::{ZX_LOG_FLAGS_MASK, dlog_header_t, dlog_record_t};
-pub const DLOG_MAX_RECORD: usize = debuglog_types::DLOG_MAX_RECORD as usize;
-pub const DLOG_MAX_DATA: usize = debuglog_types::DLOG_MAX_DATA as usize;
+pub const DLOG_MAX_RECORD: usize = debuglog_types::DLOG_MAX_RECORD;
+pub const DLOG_MAX_DATA: usize = debuglog_types::DLOG_MAX_DATA;
 
 #[allow(improper_ctypes)]
 unsafe extern "C" {
